% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processbadnumbers.R
\name{processbadnumbers}
\alias{processbadnumbers}
\title{Process Bad Numbers}
\usage{
processbadnumbers(filelocation)
}
\arguments{
\item{filelocation}{a character path to a file}
}
\value{
a data frame with one row per number that needs to be inactivated
}
\description{
Process Bad Numbers
}
