% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processconfirmationupdates.R
\name{processconfirmationupdates}
\alias{processconfirmationupdates}
\title{Process Updates Made During Confirmation}
\usage{
processconfirmationupdates(filename)
}
\arguments{
\item{filename}{a character path to a file}
}
\value{
a list of two data frames: newaddress = people with new addresses; newemail = people with new email addresses
}
\description{
Process Updates Made During Confirmation
}
