% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processphonathon.R
\name{processphonathon}
\alias{processphonathon}
\title{Process Phonathon Reports}
\usage{
processphonathon(directory = "data", startdate = Sys.time(),
  updatetypes = c("badnumbers", "wrongnumbers", "confirmations", "updates"))
}
\arguments{
\item{updatetypes}{}
}
\description{
Process Phonathon Reports
}
