% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processupdates.R
\name{processudpates}
\alias{processudpates}
\title{Process Update Files}
\usage{
processudpates(filename)
}
\arguments{
\item{filename}{a character path to a file}
}
\value{
a list of four data frames: newaddress = people with new addresses; newemail = people with new email addresses; newjob = people with new jobs; newphone = new phone numbers
}
\description{
Process Update Files
}
