% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processwrongnumbers.R
\name{processwrongnumbers}
\alias{processwrongnumbers}
\title{Process Wrong Number Reports}
\usage{
processwrongnumbers(filename)
}
\arguments{
\item{filename}{a character path to a file}
}
\value{
a list of three data frames: wrong = wrong numbers; exclusions = NOC and NPS; deceased = folks to mark as deceased
}
\description{
Process Wrong Number Reports
}
