% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processphonathon.R
\name{processphonathon}
\alias{processphonathon}
\title{Process Phonathon Reports}
\usage{
processphonathon(directory = "data",
  saveto = str_replace_all(paste0("output/phonathonupdates-", Sys.time(),
  ".csv"), ":| ", "-"), startdate = Sys.time(),
  updatetypes = c("badnumbers", "wrongnumbers", "confirmations", "updates"))
}
\arguments{
\item{directory}{the directory path where update files are stored. Defaults to getwd()/data}

\item{saveto}{the file path where you want to save the file. Use NA to bypass saving the file}

\item{startdate}{a date or time. Files modified within 10 hours or earlier than this time will be processed. Defaults to now.}

\item{updatetypes}{a character vector listing which types of updates should be processed.}
}
\value{
a single data frame with all processed updates bound together
}
\description{
Process Phonathon Reports
}
